SyncInfo as interface #1053
